Allow fractional values for strokes and added serializer info

+++ src/PixiEditor.ChangeableDocument/Changeables/Graph/Interfaces/IReadOnlyShapeVectorData.cs
@@ -9,7 +9,7 @@ public interface IReadOnlyShapeVectorData
     public Matrix3X3 TransformationMatrix { get; }
     public Color StrokeColor { get; }
     public Color FillColor { get; }
-    public int StrokeWidth { get; }
+    public float StrokeWidth { get; }
     public RectD GeometryAABB { get; }
     public RectD TransformedAABB { get; }
     public ShapeCorners TransformationCorners { get; }
